Restructure animal and dispatch React apps.
