+++ frontend/src/dispatch/DispatchResolutionForm.js
@@ -21,113 +21,10 @@ import {
 import * as Yup from 'yup';
 import Moment from 'react-moment';
 import Header from '../components/Header';
-import { Checkbox, DateTimePicker, DropDown, TextInput } from '.././components/Form';
+import { Checkbox, DateTimePicker, DropDown, TextInput } from '../components/Form';
 import { dispatchStatusChoices } from '../animals/constants';
 
-export const EvacTeamMemberForm = () => {
-
-  // Track whether or not to add another evac team member after saving.
-  const [addAnother, setAddAnother] = useState(false);
-  // Regex validators.
-  const phoneRegex = /^[0-9]{10}$/
-
-  return (
-    <Formik
-      initialValues={{
-        first_name: '',
-        last_name: '',
-        phone: '',
-        agency_id: '',
-      }}
-      validationSchema={Yup.object({
-        first_name: Yup.string()
-          .max(50, 'Must be 50 characters or less')
-          .required('Required'),
-        last_name: Yup.string()
-          .max(50, 'Must be 50 characters or less')
-          .required('Required'),
-        phone: Yup.string()
-          .matches(phoneRegex, "Phone number is not valid")
-          .required('Required'),
-        agency_id: Yup.string(),
-      })}
-      onSubmit={(values, { setSubmitting, resetForm }) => {
-        setTimeout(() => {
-          axios.post('/evac/api/evacteammember/', values)
-            .then(function () {
-              if (addAnother) {
-                resetForm();
-              }
-              else {
-                navigate('/evac');
-              }
-            })
-            .catch(error => {
-              console.log(error.response);
-            });
-          setSubmitting(false);
-        }, 500);
-      }}
-    >
-      {form => (
-        <Card border="secondary" className="mt-5">
-          <Card.Header as="h5" className="pl-3"><span style={{ cursor: 'pointer' }} onClick={() => window.history.back()} className="mr-3"><FontAwesomeIcon icon={faArrowAltCircleLeft} size="lg" inverse /></span>New Team Member</Card.Header>
-          <Card.Body>
-            <Form>
-              <FormGroup>
-                <Row>
-                  <Col>
-                    <TextInput
-                      type="text"
-                      label="First Name*"
-                      name="first_name"
-                      id="first_name"
-                    />
-                  </Col>
-                  <Col>
-                    <TextInput
-                      type="text"
-                      label="Last Name*"
-                      name="last_name"
-                      id="last_name"
-                    />
-                  </Col>
-                </Row>
-              </FormGroup>
-              <FormGroup>
-                <Row>
-                  <Col>
-                    <TextInput
-                      type="text"
-                      label="Phone*"
-                      name="phone"
-                      id="phone"
-                    />
-                  </Col>
-                  <Col>
-                    <TextInput
-                      type="text"
-                      label="Agency ID"
-                      name="agency_id"
-                      id="agency_id"
-                    />
-                  </Col>
-                </Row>
-              </FormGroup>
-            </Form>
-          </Card.Body>
-          <ButtonGroup>
-            <Button type="button" className="btn btn-primary mr-1" onClick={() => { setAddAnother(false); form.submitForm() }}>Save</Button>
-            <Button type="button" className="btn btn-success mr-1" onClick={() => { setAddAnother(true); form.submitForm() }}>Add Another</Button>
-            <Link className="btn btn-secondary" href="/evac">Cancel</Link>
-          </ButtonGroup>
-        </Card>
-      )}
-    </Formik>
-  );
-};
-
-export function EvacResolution({ id }) {
+function DispatchResolutionForm({ id }) {
 
   // Initial animal data.
   const [data, setData] = useState({
@@ -427,103 +324,4 @@ export function EvacResolution({ id }) {
   )
 }
 
-export const VisitNoteForm = ({ id }) => {
-
-    const [data, setData] = useState({
-      date_completed: '',
-      notes: '',
-      service_request: null,
-      evac_assignment: null,
-      address: '',
-      forced_entry: false,
-    })
-
-  useEffect(() => {
-    let source = axios.CancelToken.source();
-    if (id) {
-      const fetchVisitNote = async () => {
-        // Fetch Visit Note data.
-        await axios.get('/hotline/api/visitnote/' + id + '/', {
-          cancelToken: source.token,
-        })
-          .then(response => {
-            setData(response.data);
-          })
-          .catch(error => {
-            console.log(error.response);
-          });
-      };
-      fetchVisitNote();
-    };
-    return () => {
-      source.cancel();
-    };
-  }, [id]);
-
-  return (
-    <Formik
-      initialValues={data}
-      enableReinitialize={true}
-      validationSchema={Yup.object({
-        date_completed: Yup.date(),
-        notes: Yup.string(),
-      })}
-      onSubmit={(values, { setSubmitting }) => {
-        setTimeout(() => {
-          axios.patch('/hotline/api/visitnote/' + values.id + '/', values)
-            .then(
-              navigate('/hotline/servicerequest/' + values.service_request)
-            )
-            .catch(error => {
-              console.log(error.response);
-            });
-          setSubmitting(false);
-        }, 500);
-      }}
-    >
-      {form => (
-        <Card border="secondary" className="mt-5">
-          <Card.Header as="h5" className="pl-3"><span style={{ cursor: 'pointer' }} onClick={() => window.history.back()} className="mr-3"><FontAwesomeIcon icon={faArrowAltCircleLeft} size="lg" inverse /></span>{!id ? "New" : "Update"} Visit Note - {form.values.address}</Card.Header>
-          <Card.Body>
-            <Form>
-              <FormGroup>
-                <Row>
-                  <DateTimePicker
-                    label="Date Completed"
-                    name="date_completed"
-                    id="date_completed"
-                    xs="7"
-                    clearable={false}
-                    onChange={(date, dateStr) => {
-                      form.setFieldValue("date_completed", dateStr)
-                    }}
-                    value={form.values.date_completed || null}
-                  />
-                </Row>
-                <Row className="mt-3 pl-0">
-                  <TextInput
-                    as="textarea"
-                    label="Notes"
-                    name="notes"
-                    id="notes"
-                    xs="7"
-                    rows={5}
-                  />
-                </Row>
-                <Row>
-                  <Col>
-                    <Label htmlFor="forced_entry" className="mt-2">Forced Entry</Label>
-                    <Field component={Switch} name="forced_entry" type="checkbox" color="primary" />
-                  </Col>
-                </Row>
-              </FormGroup>
-            </Form>
-          </Card.Body>
-          <ButtonGroup>
-            <Button type="button" className="btn btn-primary" onClick={() => { form.submitForm() }}>Save</Button>
-          </ButtonGroup>
-        </Card>
-      )}
-    </Formik>
-  );
-};
+export default DispatchResolutionForm;
